Iterating over content => nothing; consume grid
